Hausaufgabe_Uebung_4

+++ src/h3/H3_main.java
@@ -2,7 +2,7 @@ package h3;
 
 public class H3_main {
 	public static void main(String[] args) {
-		float celsiusFloat = (float) 14.2, fahrenheitFloat = 1;
+		float celsiusFloat = 14.2f, fahrenheitFloat = 1.0f;
 		double celsiusDouble = 1.0, fahrenheitDouble = 2.0;
 		fahrenheitFloat = (celsiusFloat * 9 / 5) + 32;
 		fahrenheitDouble = (celsiusDouble * 9 / 5) + 32;
